Handle more than 2**31 parameters on GPU models (#599)
We had some issues that prevented us from learning models with more than
2**31 parameters in one of the matrices (so ~8GB of total space). This
was because we were using int32 to represent the rows/cols - and
when multiplying together we'd overflow.

Fix by using size_t. Have manually verified I can train a model with
> 25M users and 128 factors with this change (both bpr & als).

+++ implicit/gpu/matrix.h
@@ -8,11 +8,11 @@ namespace implicit {
 namespace gpu {
 // Thin wrappers of CUDA memory: copies to from host, frees in destructor etc
 template <typename T> struct Vector {
-  Vector(int size, const T *data = NULL);
+  Vector(size_t size, const T *data = NULL);
   void to_host(T *output) const;
 
   std::shared_ptr<rmm::device_uvector<T>> storage;
-  int size;
+  size_t size;
   T *data;
 };
 
@@ -21,20 +21,20 @@ struct Matrix {
   // device (if allocate=True and data != null). If allocate=false, this assumes
   // the data is preallocated on the gpu (cupy etc) and doesn't allocate any new
   // storage
-  Matrix(int rows, int cols, float *data = NULL, bool allocate = true);
+  Matrix(size_t rows, size_t cols, float *data = NULL, bool allocate = true);
 
   // Create a new Matrix by slicing a single row from an existing one. The
   // underlying storage buffer is shared in this case.
-  Matrix(const Matrix &other, int rowid);
+  Matrix(const Matrix &other, size_t rowid);
 
   // Slice a contiguous series of rows from this Matrix. The underlying storge
   // buffer is shared here.
-  Matrix(const Matrix &other, int start_rowid, int end_rowid);
+  Matrix(const Matrix &other, size_t start_rowid, size_t end_rowid);
 
   // select a bunch of rows from this matrix. this creates a copy
   Matrix(const Matrix &other, const Vector<int> &rowids);
 
-  void resize(int rows, int cols);
+  void resize(size_t rows, size_t cols);
   void assign_rows(const Vector<int> &rowids, const Matrix &other);
 
   Matrix() : rows(0), cols(0), data(NULL) {}
@@ -42,7 +42,7 @@ struct Matrix {
   // Copy the Matrix to host memory.
   void to_host(float *output) const;
 
-  int rows, cols;
+  size_t rows, cols;
   float *data;
 
   std::shared_ptr<rmm::device_uvector<float>> storage;
